Partially implement learning new spells
The learning itself works, but the simulation of it in rollouts does not
fully take into account all effects, like "read ahead tax" and "tax count".
Also the simulated id of a newly learned spell is incorrect.

Nevertheless, the current solution achieved promotion to 795th place in the
bronze league.